refactor: split executor.rs, extract fork module, remove synthesize field
Split the 3700-line executor.rs monolith into executor/{mod, docker, wave,
helpers, local}.rs. Extract fork data model (ForkManifest, merge_directions,
fork_worktree_path, cleanup) from lf/commands/flow.rs into engine/fork.rs.
Remove the configurable synthesize field from forks — always use "synthesize".
